+++ apps/ui-dev/CHANGELOG.md
@@ -1,5 +1,11 @@
 # ui-dev
 
+## 1.0.25
+
+### Patch Changes
+
+- support TreeViewProvider
+
 ## 1.0.24
 
 ### Patch Changes