+++ packages/design-system/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @elf-framework/design-system
 
+## 2.1.38
+
+### Patch Changes
+
+- add treeview component
+
+- Updated dependencies []:
+  - @elf-framework/design-tokens@2.1.38
+
 ## 2.1.37
 
 ### Patch Changes