+++ packages/design-tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @elf-framework/design-tokens
 
+## 2.1.38
+
+### Patch Changes
+
+- add treeview component
+
 ## 2.1.37
 
 ### Patch Changes